Added files via upload
Please note that DnD Next is a trade marked game by Wizards of the Coast.  This was a personal project to make encounter building easier and I in no way have profited (or intend to do so) aside from the convenience of this program.  Other GitHub users are welcome to view and utilize my code as they see fit, but please respect the Wizards of the Coast property and copy right.